google-adk-apps/newhire_onboarding_assistant/requirements.txt
# ADK Core
google-adk>=0.5.0

# Environment and Configuration
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.0

mcp
deprecated
toolbox-core

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0


# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0

# Logging and Monitoring
structlog>=23.0.0
